Split http module into ws and http
